:wrench:  export enhanced resolver
